tools/deploy/requirements.txt
# Core dependencies
requests>=2.28.0

# Google APIs (optional - for Analytics and Search Console)
google-analytics-admin>=0.22.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
